Add util::string_view
This is basically a very minimal C++17 std::string_view in a non-templated
char-only form. It can't be a 100% drop in substitute, but it'll make do before
we are able to switch.

While we don't do much string manipulation at runtime it's still nice to not
pay the compile-time, runtime and binary size cost of std::string where not
strictly needed.