Get new report configuration into json report, and add new json3 variation for profileset.
Move profileset report option out of profileset private code into the json report. This required making some profileset content publicly available, but is also more in line with the other reports.

The main change to the profileset json report is to always make metrics available as an array, instead of favouring the front metric.