Try to automatically handle "hybrid primary stat" buffs
It looks like Apply Buff A_MOD_STAT with Misc Effect == -2 is the indicator for
the hybrid primary stat buffs. Add the check in the relevant places.

Should be a no-op currently as it looks like we already handle all of the
relevant "hybrid" buffs manually.

"Found" by and fixes ubsan report:

  engine/item/special_effect.cpp:132:29: runtime error: load of value 4294967295, which is not a valid value for type 'stat_e'
    #0 0x7cef854 in (anonymous namespace)::stat_buff_type(spelleffect_data_t const&) /engine/item/special_effect.cpp:132:29