Enemy Add: fix a possible nan -> int conversion
Then the function is called the moment the add is spawned the very first ttm
calculation collapses to `0 / 0`. This is a nan. Which gets propagated all the
way to the timespan construction and results in UB.

Fix it up by explicitly handling this case.

While at it simplify (and speed up) the general case a bit.

Fixes ubsan report:

  engine/./sc_timespan.hpp:76:45: runtime error: -nan is outside the range of representable values of type 'long'
      #0 0x552568c in timespan_adl_barrier::timespan_t timespan_adl_barrier::timespan_t::from_seconds<double, void>(double) /engine/./sc_timespan.hpp:76:45
      #1 0x662a389 in (anonymous namespace)::add_t::time_to_percent(double) const /engine/class_modules/sc_enemy.cpp:941:14