[Hunter] simplify summon_pet action
Drop the 'name' option. The action now always summons the pet specified in the
'summon_pet' player-scope option (or not if it's 'disabled'). Due to the way
the module is setup they had to match anyway, you couldn't summon another pet
through `summon_pet`.

Removes one place where users could shoot themselves in the fot so to speak.