GH Actions: Add ccache caching
Our typical types of commits that touch c++ are:
 * class modules (both logic & apl)
 * spell data

Both of them touch only a small handful of files, "core" stuff is usually
pretty stable. So we can save a bunch of time caching compilation artifacts
with ccache.

In tests a single file change run with a "hot" cache takes ~1/3 of the time for
a "cold" run. There is usually no slowdown compared to no-ccache (sometimes the
cache actions decides to take a minute instead of a typical ~4 seconds /shrug).

The caches are split between clang & gcc builds to minimize the risks of there
being any problems. In case there is any problem with the cache state (there
should be none, but...) there is a global env var 'ccache-generation' which
could be incremented to force invalidation.

Small comment on the interaction with the cache action implementation.
This is a bit wasteful on the space available for cache actions. The cache
snapshots are immutable and uploaded only if there is no hit for the cache key.
As we *need* to always upload a full snapshot on every CI run for ccache to
work properly we end up filling up the cache storage with never used snapshots.
While a bit wasteful this should be fine, the cache is supposed to work as lru.
Ideally the cache action would support an "always overwrite" mode.