Arms APL Update (Fixed missing commas)
Update to the Arms APL. Very large DPS increases with Rend and Massacre, and a smaller DPS increase to Fervor of Battle. Changes to better account for a few different Major Essence choices, leading to a DPS increase for Condensed Life Force and Blood of the Enemy. Updated a few racial lines for the better. Most of these changes are larger increases the more Expedient you have.

Mostly comprehensive change list:

Rewrote precombat lines for Memory of Lucid Dreams and Guardian of Azeroth to account for shorter fight timings, leading to large dps increases for 2 minute and 3 minute fights.
Added conditionals to force certain potions for precombat based on essence choice.
Updated the second potion line.
Updated the usage of Ashvane's Razor Coral.
Updated Blood of the Enemy, Guardian of Azeroth, and Memory of Lucid Dreams lines.
Added Rend to the Execute section of the APL as well as changed its priority in the Single Target section.
Changed some conditionals for casting Execute.
Changed some conditionals for casting Whirlwind.
5 minute APL comparison sim with super high end gear Lucid - https://www.raidbots.com/simbot/report/eCUidknezTcLq8vc4GETpk
5 minute APL comparison sim with T25 Arms Mythic default Lucid - https://www.raidbots.com/simbot/report/tgPA3GebckYJ3nSDeZEL8U
3 minute APL comparison sim with T25 Arms Mythic default Lucid - https://www.raidbots.com/simbot/report/aYxc3zuTvnyKhtmgPTL4x9
2 minute APL comparison sim with T25 Arms Mythic default CLF - https://www.raidbots.com/simbot/report/79bdFNoXx3wfBAmkTFyyy