Refactor action_sequence_data_t a bit
Specifically replace an ad-hoc pair of an object + vector of doubles with a
dedicated struct. Turns out we don't really need the vector in the first place
as it has 2 elements at most. We also can not "value erase" everything into
doubles.

While at it fold the 2 copy-pasted constructors into a single one and use
inheriting.