Refactor report color object a bit
Reduces binary size by ~45kB.
Mostly be eliminating all global constructors for "literal" color constants
from all TUs including `color.hpp` (both directly and indirectly through
`simulationcraft.hpp`). We end up with only 2 constants in the binary for the 2
colors which are odr-used, everything else is gone.

This begs for C++17 inline vars.

While at it remove streams support (use fmtlib instead), change the alpha to a
float (to reduce the object size) and switch to pass-by-value (it's now the
same size as a pointer on 64b and it reduces the amount of odr usage).