Util: handroll std::tolower()
There is nothing special about it. Especially considering we only use it for
"low" ascii (it would actually fail/assert as used for "high" ascii on
platforms where char is signed).

Handrolling it lets the compiler constant fold some of the logic when one of
the chars is a compile-time constants (which happens for
`util::str_compare_ci()` with string literals).