[DBC] switch to span spell_data_t accessors
For effects/labels/powers/drivers. Instead of going directly through the
pointer to the underlying vector in the places where we don't need it.

When/if we change the way linking is done we will most definitely not have
vectors here anymore. Nor something that will behave as a vector. Change it now
as a preparation to a possible future rework.

It's also a much nicer api to work with (like when you simply need to iterate
over the whole list).