[DBC] improve items stats deduplication
Curently we dedup only stat lists that fully match (as we use them as keys in a
dictionary). Improve that a bit by using the full "flat" stats list as a
dictionary (in data compression terms).

While at it factor out all of the logic into a helper class as it's a bit
unwieldy atm.