[DBC] separate out item effects from item data
This is a bit more complicated than stats as the effects are accessed both
through the `dbc_item_data_t` (which is immutable) and through
`parsed_item_data_t` (where effects can come from wowhead/bcp_api/item
bonuses).

To minimize potential problems with effects getting out of sync (in case a
`parsed_item_data_t` is accessed through a `dbc_item_data_t` pointer) we use an
inner pointer for effects and forbid copying of `dbc_item_data_t`. Pretty sure
the problem is a bit theoretical and maybe there is a more elegant solution,
comments welcome.

We also can't deduplicate as much - the effects have indexes (so we can't sort
them) and are not always "contiguous" (there are "holes"). We *seem* to care
about that. But we still reduce `6439` total effects to 4171 unique "effect
records". The comparison is a bit off as the latter also includes 470 'nil'
effects where client data has "holes".

Reasons for not using `item_effect_t`:
 * exporting the effects as part of ItemEffectGenerator is much more
   complicated - we would have to store per item effect id arrays and then do
   some kind of runtime linking in init (which would require memory
   allocations and making otem data non-const); this seems pretty suboptimal
   to me

 * exporting `item_effect_t`s as part of ItemDataGenerator simply wastes more
   space as we can't deduplicate the effects - `item_effect_t` contains an
   (id, index) tuple which is pretty much unique (for comparison, 6439
   `item_effect_t`s is ~176KiB + we would need either 470 "nils" (~12.8KiB) or
   6909 pointers (~54KiB))

In the end it's both more simple and takes less space to use a "custom" struct
that has only what we need (and uses smaller types for some fields) storing it
in a flat array that is easier to deduplicate.

  nuohep@r2d2:~/dev/simc> bloaty -d symbols -n 5 build-optimized/simc -- build-optimized/simc.shl
      FILE SIZE        VM SIZE
   --------------  --------------
    [NEW] +65.2Ki  [NEW] +65.2Ki    __item_effects_data
    +0.0% +1.39Ki  +0.0% +1.12Ki    [753 Others]
    [NEW]    +613  [NEW]    +472    parsed_item_data_t::add_effect()
   -19.1%    -626 -19.6%    -626    enchant::initialize_relic()
   -40.0%    -633 -42.2%    -633    (anonymous namespace)::bfa::init_punchcard()
   -41.4% -6.22Mi -41.4% -6.22Mi    __item_data
   -14.3% -6.16Mi -15.1% -6.16Mi    TOTAL