Touch up static containers for gcc < 7
Turns out gcc < 7 don't like current implementation of constexpr sort
combined with pairs. Specifically it doesn't like the assignments.
Workaround that by replacing the current algorithm with a swap-based
version combined with a specialization of swap for pairs.

On the flip side we can replace the custom pair with `std::pair`.