Export Soulbind trees abilities
A bit convoluted, not sure if there is a more direct way, couldn't find
one.

Pretty sure we'll need some of the garrison talent tree junk for
supporting soulbinds anyway (don't have access to beta so can't check
what we have access to in the client for exporting through /simc).