[DBC] export a single cast_time field
Now that we dropped export of always-zero cast-time related fields its
become pretty obvious that we need a single cast_time field as the "min"
is never used in the sim.

NB. All of the spells we export have cast_min == cast_max. Interestingly
there are only 29 different values for cast time.