Implement player_t::find_runeforge_legendary( name )
Returns item_runeforge_t wrapper object holding the client data entry
for the effect, and the item associated with it. If no runeforge is
found, returns an empty wrapper object. Note that the code currently
does not perform any checks on where the effect is applied to (based on
a bonus id).

The name is an exact match to one of the RuneforgeLegendaryAbility names
(see engine/dbc/generated/item_runeforge.inc).

operator -> is overriden to return the associated spell data, or
spell_data_t::not_found(). Note that this spell data is the one
associated with the item effect, not the spell data associated with the
RuneforgeLegendaryAbility client data file.