Action: automatically parse max targets from spell data
Now that we export it, lets use it.

Apply only when it's -1 (which is unlimited targets) or when it's >1 (1
is usually set for "limit 1" spells and is also accompanied by a flag in
spell attributes, so we skip that).

Should be a noop for spells currently manually set to aoe.