Action: automatically parse some spell data attributes
This effectively moves attribute parsing from `proc_action_t` to the
base action.

One caveat is `may_crit` parsing. Turns out we rely on the false default
in a lot of places for non-direct damage actions. As the attribute
*disables* critting, parsing it unconditionally would flag a lot of
actions as `may_crit=true`. Not only this changes sim results, it also
seems a bit wasteful - more state being collected for action execution,
more rng calls, etc.
So the parsing is a bit conservative and relies on the spell having at
least one "direct damage" effect.

At the same time proc actions rely on the parsing being unconditional,
so its left there as-is.

No change for a deterministic CI.simc run.