Implement player_t::find_rank_spell( name, desc )
Finds the rank spell (rank string defined in desc parameter) for the
actor and checks that they are allowed to use it. If not,
spell_data_t::not_found() is returned. Note that this also works for
"specialization" rank spells (which you can currently also find by using
find_specialization_spell( name, desc ).