[DBC] Scrub "non-Rank" strings from Specialization Spell export
This is done to align the export file with the semantics of
find_specialization_spell, where the name + desc variant matches entries
that have both name and desc fields filled, and the name variant, that
is intended to match the fields without any rank string.

This is not really an optimal solution, and people are encouraged to use
player_t::find_rank_spell to find their rank passives for spells,
however the intention here is to allow them to use
find_specialization_spell as well, if they so choose.