Spell Info: fix printing of unknown specs
Now that most of our enums gained formatting support `util::to_string()`
on a "supported" enum will return the symbloc name instead of the
integer representation. Which is not always what we want.

Explicitly cast to an integer here to het the numeric value. While at it
refactor surrounding code to be more readable.