[Hunter] add Resonating Arrow
Initial simplified implementation. Simple buff on the player. It does
mean that it affects all enemies present for the duration.

Realistically it should match most of the sim setups. The only case I
can come up with where that would fail is
`distance_targeting_enabled=true` *and* enemies placed too far apart from
each other.

As a bonus point it makes apls easier and is much faster than something
completely matching in-game behaviour (with debuffs applied to enemies
in the aoe zone).