Add player_t::find_covenant_spell( name )
Returns the covenant spell data for a specific spell, if the actor has
the covenant enabled (and is of the correct class). Returns
spell_data_t::not_found if the covenant is not enabled on the actor, and
spell_data_t::nil if the parameter passed is not a valid covenant
ability name.