Recheck util::to_unsigned usages, and replace some with to_unsigned_ignore_error
This restores the old behaviour of just returning 0 for a string that could not be parsed in a certain places which are only interested in an optional value, and which skip further processing if the number could not be parsed that way.

Also add more detailed exception messages for conversion errors. This gives more detailed error message on init errors to the user, instead of just saying "stoi" or something like that.