Micro-optimize sample data w/ min max
First, we don't need the `found` boolean. The observation is that our
default min and max are chosen such that min > max. If we get at least
one sample it necessarily has to update one of them so their relation
becomes the "proper" min <= max.

Second, convert the setters into ternaries. On x64 this lets the
compiler update min/max with a single instruction (`minsd`/maxsd`)
without any branching.