[Hunter] preliminary Master Marksman
Implement 2 different behaviours:
* current in-game behaviour
* Piercing Shots -like rolling dot

The former is the default, the latter can be activated by
`hunter.rolling_master_marksman` option.

The way it works in-game is not thoroughly tested and the implementation
is mostly a hack. The hope is that it will get fixed (changed to a
rolling dot) shortly. The talent does not make any sense as a
non-rolling dot anyway, especially considering Rapid Fire / Barrage.

F