[Stats] Option to disable counting dot extend/reduce as refreshes
An action is not displayed in the DTEP chart on the HTML report if refreshes * 4 > executes. For those actions where there are many small adjustments that will not cause it to look unwieldy when presented on the chart, an optional last argument to dot_t::extend_duration and dot::reduce_duration, when set to false, will not increment the refresh counter, and thus allow it to be displayed on the DTEP chart.