[DBC] experimental per-player spell overrides
Add an `override.player.spell_data=` option which is the same as
`override.spell_data=` but scoped only to the current player.

Big caveat is that the option has a non-obvious "position dependence" to
the sim-wide override one (due to the way we parse options). To not hit
any surprising behavior *all* sim wide overrides should go before any
actors.

A bit beta quality as I'm not entirely sure about the "ownership" model
& the behavior re sim overrides.