[Hunter] Wild Spirits *should* affect pet damage
This is a bit preliminary without in-game testing but looking at the
spell data of Hunter's Mark & Condensed Life-Force (which do affect pet
damage) it *seems* that this is driven by either effect sub type 380 or
381. Which means that Wild Mark (the debuff applied by Wild Spirits
in-game) *should* also affect pet damage.

Depending on if it's 380 or 381 it may *also* apply to Resonating Arrow
with Enfeebled Mark conduit as it has an effect with sub type 381.