Buff: add simplified ::trigger() overloads
Most of the time when we trigger buffs with custom values it's to supply
either a custom duration or stacks + duration. Having to also pass in
value & chance, which have to be special default values, is not ideal.

Add 2 overloads that should cover the majority of custom buff triggers
without requiring needless special arguments.