[DBC] try to properly handle hotfixed entries
We currently have a problem where we don't filter out duplicate entries
from the parent & reference caches.

In the DataGenerators this manifests in "duplicate" entries being
returned for `children()` and `child_refs()` (first for the db2 record,
second for the hotfixed record) and `child()` returning the non-hotfixed
record (and complaining that there is more than one).

Fix that up by storing the references in a dict and converting them to a
tuple on first access.