Fix an issue with spell_targets expression
If the spell_targets expression was used with
distance_targeting_enabled, the state of the "targeted spell" (action)
could be left in modified state.

In conjunction with channeled spells, this can cause a crash in the dot
system for the channeled dot due to the current target (broken by
spell_targets expression) not having an active dot on it. More
generally, an action's target might change unpredictably, causing
actions to execute on the wrong target.

This commit adjusts the spell_targets expression so that it restores the
original target of the action, and also invalidates the target cache
explictly after every evaluation.