[shaman] Update PreRaid generator to prepare for Shadowlands
Currently just an export of the item level 138 gear from template
characters on the beta servers, with talents reset to first column
and character level set to 60.

This is intended to be a placeholder until the class module is in
a good state in order to determine the actual best pre-raid gear
and rules for Shadowlands pre-raid are set.