Update Visual Studio version compile-time check
We require Visual Studio 2017 or newer now for C++14 support. Be a bit
conservative and check only for the initial RTW version. Pretty sure we
actually require something newer than that (we build with 15.9 in CI, so
support for older versions depends mostly on devs building it on their
own).

While at it remove a check for VS < 13 as it's not relevant anymore.