[Hunter] tweak Lock and Load handling
Fix an issue where `action_t::ready()` does not consider Aimed Shot with
Lock and Load up to cost 0 focus. This effectively means that the robo
hunter still needs 35 (or whatever the current AiS cost is) focus just
to consider casting AiS even with LnL up (it *will* be free).