BCP API: replace static map with a plain array
Not sure why but this generates intermittent compile time errors with
gcc6. It looks like some obscure bug in the constexpr evaluator that's
hard to workaround (I can't hit it locally).

As the "map" is small enough and not on a fast path replace it with a
simple array and a linear search.