Expressions: move FLOOR/CEIL handling out of the lexer
FLOOR/CEIL are a bit special in that they really aren't "tokens"
(lexemes) in the classical sense. The reason we do them as tokens seems
mostly to be because it was just easier to do them this way and to make
expressions (which are basically AST nodes) have a unique token type for
them.

In reality they both are "pseudo-functions" - a special case of a
generic single argument function - a [TOK_STR, TOK_LPAR] sequence with a
"known" value for TOK_STR (which is actually an "identifier"
(imaginary TOK_IDENT), really).

So handling them in the lexer does not really make sense. It would be
much more apparent if we did single-pass parsing (instead of ~3 passes
we do now).