[shaman] WIP Elemental Equilibrium
After further testing on beta, this legendary is in a really funky state
and is currently generating the buff both on-cast and on-impact, and
some spells are consuming buffs they're not supposed to.

I'm going to leave this as a WIP draft until a new build arrives
that fixes all the crazy bugs with this legendary.