Fix cases where `use_item` action will ignore the effect's `custom_buff`
When an effect has a custom_buff, the expected behavior is that it takes precedence over automatically parsed/created buffs. In init() for action_t::use_item_t, during effect parsing/buff creation, a name search is made for a buff matching the name of the effect, without checking to see if custom_buff exists. This breaks expected behavior, and also is redundant as special_effect_t::create_buff() performs the same name search but with a check for custom_buff.