Disable protection paladin and protection warrior.
Profiles are not yet in a working condition for shadowlands. Deactivate for now so that it does not get forgotten before launch.

Classes can still be run with allow_experimental_specializations=1 to develop them until they are ready to be fully activated again.

Fix Warrior Protection deep wounds out of bounds effect access.