Tests: switch to custom python runner from bats (#5382)
Out current tests are a bit arcane to setup and run because they use the
bats testing framework. While it's good and tested we use it only as
basically a glorified for loop, even calling to python for talent string
generation.

As we already use python for some parts of testing we may as well fully
switch to it. While it's true that it's another piece of code we have to
maintain, it's actually less code than we had with bats and has a bunch
of benefits:
* an actual language for setting up & running tests
* one less obscure dependency
* possibility to easily run tests locally

The "framework" is pretty limited & simple for now but seems to fully
replace what we have before. The main idea is that you simply add tests
into a list (with a possiblity to group them) and then simply run them.