Add trinket tests.
Use argparse for run.py since I was not happy with the silent fail when invoking it without arguments.

simc-support is a helper library written by Bloodmallet which contains various game data as json and is easily accessible with python. It is a outside dependency, but I thought it was worth the effort. Can be disabled with --no-trinkets option.