add `interpolated_fight_remains` expression
`interpolated_fight_remains` expression returns a simple linear interpolation between time remaining based on the sim
scope max_time setting and time ramining based on iteration-specfic fight length after variance is applied to max_time.

Currently `fight_remains` uses the iteration-specific fight length, which is max_time adjusted for the variance.
Generally, this is information that is not known to the player at the start of the encounter, where max_time is the
expected length of combat. As the encounter progresses, the player's estimation of time remaining converges towards
the iteration-specific fight length, converging at the end of the fight.