[Buffs] implement _initial_stack variable & related functions
When a buff is trigger()'d or execute()'d with the default parameter of
`stack == -1` it will appropriately increment() a number of stacks equal
to the _intial_stack value (for reverse buffs, behavior is unchanged and
will call decrement(1)).

`buff_t::set_initial_stack( int )` to set the _initial_stack value.
`buff_t::modify_initial_stack( int )` to add to the current _initial_stack.