[Paladin] Add proc tracking for avengers' shield cooldown reset mechanics
Add a proc object and its "wasted proc" partner object for Grand Crusader (GC), Moment of Glory (MoG)and Holy Avenger's Engraved Sigil (HAES).

Current behaviour is that a HAES proc during MoG is a waste of the HAES proc and that casting MoG or getting a GC proc while Avenger's Shield is off cooldown is a waste.