[Item] Anima Field Emitter
By default, allows full (8) seconds of haste rating buff.

Two new sim-scope options are defined to control the duration of the
hast rating buff.
  * shadowlands.anima_field_emitter_mean: The mean of the normally
  distributed variable that controls the duration of the haste buff.
  Defaults to "always use 8 seconds".
  * shadowlands.anima_field_emitter_stddev: The standard deviation of
  the normally distributed variable that controls the duration of the
  haste buff. Defaults to 0.

When using the random variable to control the duration of the buff, it
will always be clamped between 0 and 8 seconds. If the random variable
results in a negative value, the buff will not be triggered.