[Brewmaster] Invoke Niuzao R2 + Misc fixes (#5585)
* implement Invoke Niuzao, Rank 2 (Hot Trub effect)

if Nizzy ONLY cared about purifies during his effect, then we could
track this with a single double. However, because he ALSO sees casts
from up to 6s before being cast, we need to a bit more tracking.

* implemented rank 2 of purifying brew

just adds a charge. this got missed in the BFA->SL transition. the
charge used to be baseline

* walk with the ox now only applies to stomp

* fixed scalding brew implementation

the previous implementation had the correct calculation, but applied the
calculation too late to take effect

* niuzao's stomp splits damage

* remove faeline stomp bof interaction

* switched invoke niuzao over to buff implementation