[Gear] Allow precombat Soul Igniter
This trinket has a 15 second duration buff on the player and does not do anything harmful until this buff expires. To model this in simc, simply treat the first expire after a precombat cast of Soul Igniter as a full duration (15 second) expiration. Individual APLs can use cancel_buff,name=soul_ignition to decide when to cancel it, as there are some reasons to do it later than at 0 seconds.