Initial implementation of core temporary enchant system.
- New player-scope option "temporary_enchant" added that takes a list of
tempoary enchants as an argument in the format 'slot:enchant_name' using
'/'-character as element separator.

Some notes on the implementation:
- Changes item initialization slightly; now requires that for any
"special effect", the item must be "active" for special effects to be
initialized
- player_t::default_temporary_enchant can be overridden in the class
modules to define the default temporary enchants for each profile
- Simc implementation allows "invalid" temporary enchants on items, in
the sense that no inventory type or ilevel checks are made. Item level
checks put a nag on error stream about failing the min/max ilevel check.