Implement simple "stuck sim" detection
Based on keeping track of time progression in the simulator. If time has
not moved for a number of consecutive executed events, deem the
simulator (thread) stuck, and initiate cancelation of the work.

Attempts to tell to the user what actor causes the stuck state, but in
some cases (namely, multi-actor simulations) this is impossible.

The implementation considers (500 * active actors) consecutive events
where time does not progress as a qualified "stuck state".