Make action_t::parse_effect_data non virtual
Since the function is called in the action_t constructor it should not be virtual.

Heal_t is the only place overriding this, and already had to manually call it again on all effects. Just rename it there to make its purpose clearer, and and make it non-virtual in action_t.