[Spell Query] hide spells for common labels
Some labels, such as label 16, are present on a large percentage of spells that we generate. Showing these on all spells adds a lot of extra lines to the output and currently does not provide much useful information. The long list of spells that gets added in spell effect output can just be replaced with a simple description of the label.