Use range contains and any_of
Simplifies code where the intention is only to check if a element is contained in a range or not. No need for that special compare against the end iterator.

Personally I do like contains as a name better than any_of, but the latter is I guess technically more correct and it matches the new STL algorithm std::any_of