Util: static_map - fixup infinite recursion
gcc10/clang12 in C++17 mode correctly point out that these operators are
written horribly and result in an infinite recursion - calling itself to
do the conversion instead of invoking `view_type` constructor (which was
the intention).

Seems like it's either a change in C++17 rules for overload resolution
in templates, a result of a change due to a defect report or simply
compilers starting doing it differently to align it with non-template
contexts (where it would be an obvious bug all the way back).

Fix it up by explicitly casting the objects to their base types to
properly invoke the desired constructors.