[Soulbinds] implement Newfound Resolve
shadowlands.newfound_resolve_success_chance sets the chance that the player faces their Doubt at all.

shadowlands.newfound_resolve_default_delay and shadowlands.newfound_resolve_delay_relstddev define the Gaussian distribution of when the player will face their Doubt. This is clamped between around 2.9 seconds (this is the fastest you can get the buff even if you are already facing your Doubt) and 11 seconds (the longest you can wait before it would despawn).

If the newfound_resolve action is present in an APL, that will disable the delay options (not the success chance option) and specify when the Newfound Resolve buffs are picked up. However, if the Trial of Doubt debuff expires, the Newfound Resolve buff will still be given even if the APL fails to use the action. This means that the conditions for this action only need to specify situations where the player would want to gain the buff early.