[Druid] adjust convoke for celestial spirits (WIP)
As of commit time, celestial spirits is bugged and does not proc exceptional spells as expected for a 2 deck rng. Given this along with the obscure nature of the spell (server side script, no combat log events) as well as the weakness of the legendary, testing of these values are very limited and of low confidence.