[Mage] implement Frostbolt proc BLP
There appears to be a bad luck protection (BLP) system for Frostbolt to guarantee either an FoF or BF proc after an unlucky streak of casts. Note that because there are two proc types, there does not seem to be a meaningful way to play around this system. However, it does have an effect on when procs occur and has a significant but small impact on overall DPS in a few cases.