Do not check if C compiler works with cmake.
We are not really using it anyway, and it seems something might not be correctly set working for the gcc-7 job (maybe related to some system libraries).

Switch cmake version to 3.10.
QT 6 recommends 3.16.
Ubuntu 18.04 uses 3.10, our linode debian server has 3.12.

And since we have 3.21 currently running in the pipeline, I'll try getting everything running with that as the max version. We'll see.