Add cmake install target and try to cleanup paths. (#6185)
* Add cmake install target and try to cleanup paths.

- Normal install on UNIX to bin and share.
- Flat install on Windows, can be used for creating package for nightly.

Try to use the paths by default in Simc, with less additional build variable setup. Setting SC_SHARED_PATH should still work though.

Cmake install path setup can of course be improved, and maybe it can also be used as a starting point for the macos bundle, not sure.

* Check if cmake minimum version is causing gcc7 configure fail.

* Do not check if C compiler works with cmake.

We are not really using it anyway, and it seems something might not be correctly set working for the gcc-7 job (maybe related to some system libraries).

Switch cmake version to 3.10.
QT 6 recommends 3.16.
Ubuntu 18.04 uses 3.10, our linode debian server has 3.12.

And since we have 3.21 currently running in the pipeline, I'll try getting everything running with that as the max version. We'll see.

* Try setting cmake project to CXX and maybe cmake will not check for a C compiler.