[dbc] fix some incorrect hotfix outputs
For some reason, Blizzard sometimes seems to push hotfix entries with HotfixType.REMOVED that reference records that do not exist. When this happens, the id will not end up in replaced_ids, which would resullt in the addition of an empty record.

In the current live build (9.1.5.40966), this occurs with a HotfixType.Removed entry for SpellCategory record 1999.