Expressions: update for std::string_view transition
Simply return EOF token here without going through yield_token() as all
of the string manipulations there are meaningless for EOF.

The actual bug here is that for EOF current_len still contains the
length of the previous token, so remove_prefix() in yield_token() gets
called with a non-zero n while the input string is empty. This is UB for
std::string_view and all of the stls assert for it (in some way or
another) while our implementation of string_view simply clamps the
passed in value.