Do not set `-march=native` on darwin/arm64
The version of `clang` in XCode 13 does not support `-march=native` for
this chipset, so we exclude it from the `optimized` target by default on
that particular architecture.

This means darwin/arm64 builds may be slightly less optimized than non-darwin/arm64
builds, but that these do work out of the box without requiring extra
flags on my 2021 MacBook Pro M1 Max.