[HTML] adjust reporting of queued actions that become invalid before executing
Because APL action line execution data, as well as sample sequence, is recorded
upon the action being queued, rather than execute, it's very rarely possible
for the queued action to no longer be valid when it becomes time to execute.

This usually happens with abilities that have cooldowns and/or abilities with
dynamic adjustments to their cost.

When this happens, the action counter for HTML reporting purposes is properly
decremented, a proc is created with the name `<action name> (queue failed)`,
and if it was during the first iteration, the sample sequence entry is editted
to read `<action name> (queue failed)`.