[Balance] manually parse data for adaptive swarm heal
NOTE this is a temporary workaround as currently the heal is implemented
as a 0-damage spell on enemies. Eventually this needs to be re-implemented
using a different method than standard action_t/dot_t as it is possible to
have multiple healing swarm active at the same time amongst your group members,
even during a single target encounter.