Expressions: Simplify is_constant interface (#6234)
No longer require that is_constant returns the constant value as an out parameter. Instead, assume that for a expression that is promised to be constant, we can just call evaluate() and get the constant value.
The constant checks are only performed in specific core code when optimizing expressions and don't mind that extra step, while the implemented of a potentially const expression now has a much simpler job for providing info whether the expression is constant or not.

Also removes any potential bugs where the out parameter of constant could potentially not be equal to the result of a evaluate() call on that expression.

Add some more is_constant overrides to action expressions.