Generic: refactor accumulate()
Drop all of the overloads and leave a single projection-enabled one. We
don't need the BinaryOp-enabled overload at this time.

Also, while at it open code it instead of dispatching through the std
variant. Lets us do several things:
- "back-port" the C++20 change of std::move()ing the accumulator
- properly handle the common "gotcha" of summing fp values (or unwanted
  casting in general):
    `accumulate( (const double[]){ .25, .75 }, 0 )` will dtrt instead of
    returing 0 like the std version would
- drop <numeric> include from the generic header

The only usage of a BinaryOperation-enabled overload can be easily
switched to the non-binary one. If we actually ever need one it's better
to implement it then, it should be called fold anyway.