[Shaman] Implement active Flame Shock cap for 9.2
Current in-game testing seems to indicate that when going over the
target cap, one of the oldest applied Flame Shock on a target will be
removed. If multiple Flame Shocks are applied on the same timestamp
(e.g., Lava Lash spread behavior for 9.2), the simulator picks a random
one to cancel.